move `fly metrics` into `fly settings analytics`
