+++ internal/command/settings/settings.go
@@ -1,0 +1,16 @@
+package settings
+
+import (
+	"github.com/spf13/cobra"
+	"github.com/superfly/flyctl/internal/command"
+)
+
+func New() *cobra.Command {
+	cmd := command.New("settings", "Manage flyctl settings", "", nil)
+
+	cmd.AddCommand(
+		newAnalytics(),
+	)
+
+	return cmd
+}
